Issue #21: Most of the refactoring is done, but I want to siwtch to using Query for the url.URL in parseTorrentRequestURI
